modules/module_rgba/requirements.txt
Pillow
numpy
redis
